requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
python-dotenv==0.18.0
requests==2.25.1
chromedriver-autoinstaller==0.6.4
selenium==4.25.0